test_type_hints: convert INFINITY to gdouble
The *INFINITY* is defined as float by default,
but according to C99 standard it still should be fine to mean
infinite in case of double as well.

+++ lib/tests/test_type_hints.c
@@ -204,7 +204,7 @@ ParameterizedTestParameters(type_hints, test_double_cast)
   static StringDoublePair string_value_pairs[] =
   {
 #ifdef INFINITY
-    {"INF",INFINITY},
+    {"INF",(gdouble)INFINITY},
 #endif
     {"1.0",1.0},
     {"1e-100000000",0.0}
